+++ lib/src/widgets/easy_dynamic_theme_btn.dart
@@ -1,10 +1,10 @@
-import 'package:dynamic_theme/dynamic_theme.dart';
 import 'package:flutter/material.dart';
+import '../../easy_dynamic_theme.dart';
 
-class DynamicThemeBtn extends StatelessWidget {
+class EasyDynamicThemeBtn extends StatelessWidget {
 
   IconData _getIcon(BuildContext context) {
-    var themeMode = DynamicTheme.of(context).themeMode;
+    var themeMode = EasyDynamicTheme.of(context).themeMode;
     return themeMode == ThemeMode.system ? Icons.brightness_auto : themeMode == ThemeMode.light ? Icons.brightness_high : Icons.brightness_4;
   }
 
@@ -12,7 +12,7 @@ class DynamicThemeBtn extends StatelessWidget {
   Widget build(BuildContext context) {
     return RawMaterialButton(
       shape: CircleBorder(),
-      onPressed: (){DynamicTheme.of(context).switchTheme();},
+      onPressed: (){EasyDynamicTheme.of(context).switchTheme();},
       child: new Padding(
         padding: const EdgeInsets.all(12.0),
         child: new Icon(_getIcon(context), color: Theme.of(context).textTheme.button.color),